requirements.txt
opencv-python>=4.8.0
mediapipe==0.10.32
protobuf>=5.26
numpy>=1.24.0
Pillow>=10.0.0
rembg>=2.0.50
flask>=3.0
qrcode[pil]>=7.4
